# src/main/resources/public/docs/openapi.yml
{"openapi":"3.0.1","info":{"title":"Welcome to the File API","description":"Test assignment: File API","contact":{"email":"kirill.malikov@mail.com"},"version":"1.0"},"servers":[{"url":"http://localhost:6011/"}],"security":[{"Basic":[]}],"paths":{"/files":{"post":{"tags":["Files"],"summary":"Upload file","operationId":"uploadFiles","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"contentType","in":"query","required":true,"schema":{"type":"string"}},{"name":"meta","in":"query","required":true,"schema":{"type":"string","description":"JSON of additional meta","example":"{\"creatorEmployeeId\":1}"}},{"name":"source","in":"query","required":true,"schema":{"type":"string","example":"timesheet, mss, hrb, ..."}},{"name":"expireTime","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["content"],"type":"object","properties":{"content":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/files/metas":{"post":{"tags":["Files"],"summary":"Get files metadata","operationId":"getMetas","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMetasRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetMetasResponse"}}}}}}},"/status":{"get":{"tags":["Status"],"summary":"Get status","operationId":"getStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseEntityMapStringObject"}}}}}}},"/file/{token}":{"get":{"tags":["Files"],"summary":"Download file","operationId":"downloadFile","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}},"delete":{"tags":["Files"],"summary":"Delete file","operationId":"deleteFile","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"GetMetasRequest":{"required":["tokens"],"type":"object","properties":{"tokens":{"type":"array","items":{"type":"string"}}}},"FileMetaResponse":{"type":"object","properties":{"token":{"type":"string","format":"uuid"},"fileName":{"type":"string"},"size":{"type":"integer","format":"int64"},"contentType":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"meta":{"type":"string"}}},"GetMetasResponse":{"type":"object","properties":{"files":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/FileMetaResponse"}}}},"ErrorMessage":{"type":"object","properties":{"message":{"type":"string"},"code":{"type":"string"}}},"ResponseEntityMapStringObject":{"type":"object","properties":{"data":{"type":"object","additionalProperties":{"type":"object"}},"errors":{"type":"array","items":{"$ref":"#/components/schemas/ErrorMessage"}},"status":{"type":"integer","format":"int32"}}}},"securitySchemes":{"Basic":{"type":"http","description":"File API uses basic HTTP authentication,\na request contains a header field in the\nform of <b>Authorization: Basic \\<credentials\\></b>,\nwhere credentials is the Base64 encoding of\nuser and password joined by a single colon <b>:</b>","name":"Authorization"}}}}